Fix: e-mail attachments problems
Based (but adapted and improved) on my old changes at sourceforge:
@see: https://sourceforge.net/u/cbello/netimplicitssl/ci/3d8ac39620390e9ff7f8c1fbaf1ab55f2d94fc4f/

I was asked to make a pull request of my changes, but I'm never did it.
Sorry.

- Small fixes in e-mail attachment headers.
- Used attachment.name property instead of FileName for set the
attachment name.
- Now we can use ignored property NameEncoding for set encoding of the
attachment name.
- Refactor constructors of MimeAttachment class. Note: ContentType
property it's initialized in the base class constructors and should not
be reassigned.

With these changes, works in the similar way as the original
System.Net.Mail does and fix #6